Pushing to Github
Hello!
I was unsure how I should submit this piece of work and couldn't figure out how to push it to github, so created a repo and then tried to push my code to it (as everyone was saying that i can just submit the link to my github repository, but i thought we had to push the code there because we had been provided a link through our classwork)
Anyway! I confused myself so i opened a new repo, pasted my work into it and will send the link to that!
I hope you're all doing well!